Added icon snooper window to show off more library functions.

+++ DataTools/Text/FriendlyTypes.cs
@@ -473,7 +473,7 @@ namespace DataTools.Text
 
         public override string ToString()
         {
-            return TextTools.PrintFriendlySpeed((long)Value);
+            return TextTools.PrintFriendlySpeed(Value);
         }
     }
